# docs/swagger.json
{"openapi":"3.0.0","paths":{"/api/inventory/v1/inventory":{"get":{"operationId":"AppController_find","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/GenericInterface"}}}}]}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"post":{"operationId":"AppController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInventoryDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/inventory/alert":{"get":{"operationId":"AppController_findProductWithAlert","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/GenericInterface"}}}}]}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/inventory/{id}":{"get":{"operationId":"AppController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"put":{"operationId":"AppController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInventoryDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"delete":{"operationId":"AppController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/inventory/ean/{ean}":{"get":{"operationId":"AppController_findByEan","parameters":[{"name":"ean","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/store":{"get":{"operationId":"AppController_findStore","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/GenericInterface"}}}}]}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"post":{"operationId":"AppController_createStore","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStoreDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/store/{id}":{"get":{"operationId":"AppController_findOneStore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"put":{"operationId":"AppController_updateStore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStoreDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"delete":{"operationId":"AppController_deleteStore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/user":{"get":{"operationId":"AppController_findUsers","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/GenericInterface"}}}}]}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"post":{"operationId":"AppController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/user/{id}":{"get":{"operationId":"AppController_findOneUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"put":{"operationId":"AppController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]},"delete":{"operationId":"AppController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/inventory/v1/login":{"post":{"operationId":"AppController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}}},"info":{"title":"Inventory Backend","description":"Inventory Backend","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ErrorDefault":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"},"error":{"type":"string"}},"required":["statusCode","message","error"]},"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"data":{"type":"array","items":{"type":"array"}},"meta":{"$ref":"#/components/schemas/PageMetaDto"}},"required":["data","meta"]},"GenericInterface":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","description","createdAt","updatedAt"]},"CreateInventoryDto":{"type":"object","properties":{"ean":{"type":"string"},"productName":{"type":"string"},"expiresIn":{"format":"date-time","type":"string"},"daysBeforeRemove":{"type":"number"},"store_id":{"type":"number"}},"required":["ean","productName","expiresIn","daysBeforeRemove","store_id"]},"UpdateInventoryDto":{"type":"object","properties":{"ean":{"type":"string"},"productName":{"type":"string"},"expiresIn":{"format":"date-time","type":"string"},"daysBeforeRemove":{"type":"number"},"store_id":{"type":"number"}}},"CreateStoreDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdateStoreDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"store_id":{"type":"number"}},"required":["username","password","store_id"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"store_id":{"type":"number"}}},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}}}